Move grd2cpt function from gridops.py to grd2cpt.py
