+++ pygmt/__init__.py
@@ -16,12 +16,13 @@ from pygmt import datasets
 from pygmt.figure import Figure
 from pygmt.filtering import blockmedian
 from pygmt.gridding import surface
-from pygmt.gridops import grd2cpt, grdcut, grdfilter
+from pygmt.gridops import grdcut, grdfilter
 from pygmt.mathops import makecpt
 from pygmt.modules import GMTDataArrayAccessor, config, grdinfo, info, which
 from pygmt.sampling import grdtrack
 from pygmt.session_management import begin as _begin
 from pygmt.session_management import end as _end
+from pygmt.src import grd2cpt
 from pygmt.x2sys import x2sys_cross, x2sys_init
 
 # Get semantic version through setuptools-scm
