+++ pygmt/tests/test_grd2cpt.py
@@ -5,11 +5,12 @@ import os
 
 import numpy as np
 import pytest
-from pygmt import Figure, grd2cpt
+from pygmt import Figure
 from pygmt.datasets import load_earth_relief
 from pygmt.exceptions import GMTInvalidInput
 from pygmt.helpers import GMTTempFile
 from pygmt.helpers.testing import check_figures_equal
+from pygmt.src.grd2cpt import grd2cpt
 
 TEST_DATA_DIR = os.path.join(os.path.dirname(__file__), "data")
 POINTS_DATA = os.path.join(TEST_DATA_DIR, "points.txt")
